+++ src/main/java/com/almostreliable/lootjs/mixin/LootItemFunctionMixin.java
@@ -1,8 +1,8 @@
 package com.almostreliable.lootjs.mixin;
 
+import com.almostreliable.lootjs.LootJS;
 import com.almostreliable.lootjs.loot.LootConditionList;
 import com.almostreliable.lootjs.loot.extension.LootItemFunctionExtension;
-import dev.latvian.mods.kubejs.util.ConsoleJS;
 import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
 import org.spongepowered.asm.mixin.Mixin;
 
@@ -13,7 +13,8 @@ public interface LootItemFunctionMixin extends LootItemFunctionExtension {
 
     @Override
     default LootItemFunction lootjs$when(Consumer<LootConditionList> consumer) {
-        ConsoleJS.SERVER.info("Non conditional loot functions are not supported yet! Added conditions will be ignored!");
+        LootJS.DEBUG_ACTION.accept(
+                "Non conditional loot functions are not supported yet! Added conditions will be ignored!");
         return (LootItemFunction) this;
     }
 }